KE2: restore basic function label construction
